# feat: comprehensive performance optimization with test suite and caching